Added code for player hurt and death (untested)
added some code in the PlayerBehaviour to serve as the basis for the hurt and death animations.
Again, I still need to make the transition states via animator and actually test it, but this should serve as a good skeleton if anyone else would like to add on/ fix this.