updated death/hurt animations
death/hurt animations seem to work, but the hurt animation seems to be cancelling out the death animation.

Also, when i did this, I edited the pillPlayer object in the hierarchy, which seemed to result in the animations only applying to scene1. I figure this should be an easy fix.